+++ phpunit.xml
@@ -6,16 +6,18 @@
 		convertWarningsToExceptions="true"
 		stopOnFailure="false">
     <testsuites>
-        <testsuite name="Test Suite">
+        <testsuite name="unit">
             <directory>./tests/unit</directory>
         </testsuite>
     </testsuites>
     <logging>
+<!--
         <log type="coverage-html" target="./tests/log/coverage.html"/>
         <log type="coverage-text" target="php://stdout" showUncoveredFiles="false"/>
         <log type="json" target="./tests/log/logfile.json"/>
         <log type="tap" target="./tests/log/logfile.tap"/>
         <log type="testdox-html" target="./tests/log/testdox.html"/>
         <log type="testdox-text" target="./tests/log/testdox.txt"/>
+-->
     </logging>
 </phpunit>
